refactor: Convert `CfnInclude` warning from Validation to Aspect
> Aspects are a way to apply an operation to all constructs in a given scope.
> The aspect could modify the constructs, such as by adding tags,
> or it could verify something about the state of the constructs,
> such as ensuring that all buckets are encrypted.
> – source: https://docs.aws.amazon.com/cdk/v2/guide/aspects.html

An Aspect is more suited for adding this annotation to a stack.

See:
  - https://github.com/guardian/cdk/pull/1286
