+++ src/aspects/__data__/cfn.yaml
