+++ src/constructs/core/stack.test.ts
@@ -1,10 +1,8 @@
-import { join } from "path";
-import { Annotations, App } from "aws-cdk-lib";
+import { App } from "aws-cdk-lib";
 import { Match, Template } from "aws-cdk-lib/assertions";
 import { Role, ServicePrincipal } from "aws-cdk-lib/aws-iam";
 import { Queue } from "aws-cdk-lib/aws-sqs";
-import { CfnInclude } from "aws-cdk-lib/cloudformation-include";
-import { ContextKeys, LibraryInfo, MetadataKeys } from "../../constants";
+import { ContextKeys, MetadataKeys } from "../../constants";
 import { GuTemplate } from "../../utils/test";
 import type { GuStackProps } from "./stack";
 import { GuStack } from "./stack";
@@ -74,17 +72,5 @@ describe("The GuStack construct", () => {
 
     const stack = new MigratingStack(new App(), "Test", { stack: "test", stage: "TEST" });
     GuTemplate.fromStack(stack).hasResourceWithLogicalId("AWS::SQS::Queue", "Notifications");
-  });
-
-  it("prompts to use GuStack.overrideLogicalId", () => {
-    const warning = jest.spyOn(Annotations.prototype, "addWarning");
-
-    const app = new App();
-    const stack = new GuStack(app, "Test", { stack: "test", stage: "TEST" });
-    new CfnInclude(stack, "Template", { templateFile: join(__dirname, "__data__", "cfn.yaml") });
-    app.synth();
-    expect(warning).toHaveBeenCalledWith(
-      `As you're migrating a YAML/JSON template to ${LibraryInfo.NAME}, be sure to check for any stateful resources! See https://github.com/guardian/cdk/blob/main/docs/stateful-resources.md.`
-    );
   });
 });
